feat: add `cosid-spring-redis` to example
